requirements_google_calendar.txt
# Google Calendar API dependencies
google-auth>=2.15.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0
pytz>=2022.7 